+++ makdo-md2docx.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python3
 # Name:         md2docx.py
 # Version:      v02 Shin-Hakushima
-# Time-stamp:   <2022.12.16-13:42:37-JST>
+# Time-stamp:   <2022.12.17-08:51:41-JST>
 
 # md2docx.py
 # Copyright (C) 2022  Seiichiro HATA
@@ -870,8 +870,11 @@ class Document:
                 docx_file = re.sub('\\.md$', '.docx', md_file)
                 self.docx_file = docx_file
         if os.path.exists(docx_file):
+            if not os.access(docx_file, os.W_OK):
+                sys.stderr.write('error: overwriting a unwritable file\n')
+                sys.exit(1)
             if os.path.getmtime(md_file) < os.path.getmtime(docx_file):
-                sys.stderr.write('error: overwriting newer file\n')
+                sys.stderr.write('error: overwriting a newer file\n')
                 sys.exit(1)
             if os.path.exists(docx_file + '~'):
                 os.remove(docx_file + '~')
